Major bug fix to ddf model fitting
Repeating the model fitting using do.call rather than eval causes the
call component of the ddf object to become the function not the call.
This means that dht lacks a lot of information about the model call…
whether there is truncation, point transects, binned data etc.
Resulting in incorrect estimates of density and abundance.